+++ opentelemetry-user-events-metrics/CHANGELOG.md
@@ -6,4 +6,5 @@
 
 ### Added
 
+- Allow exporter to have an instance of TracepointState [#953](https://github.com/open-telemetry/opentelemetry-rust/pull/953).
 - Initial Alpha implementation